corrected return value described in doc
git-svn-id: http://svn.python.org/projects/python/branches/okkoto-sizeof@63764 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Python/sysmodule.c
@@ -659,7 +659,7 @@ sys_getsizeof(PyObject *self, PyObject *args)
 }
 
 PyDoc_STRVAR(getsizeof_doc,
-"getsizeof(object) -> long\n\
+"getsizeof(object) -> int\n\
 \n\
 Return the size of object in bytes.");
